requirements.txt
# Core data science packages
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scipy>=1.7.0

# Bioinformatics
biopython>=1.79

# Statistical analysis
statsmodels>=0.13.0

# Progress bars and utilities
tqdm>=4.62.0

# For notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# For sequence alignment (if using minimap2 via python)
pysam>=0.19.0

# For molecular biology work
primer3-py>=0.6.0

# Plotting enhancements
plotly>=5.0.0

# Note: External tools required:
# - minimap2 (for sequence alignment)
# - vsearch (for sequence processing)
# - FLASH (for read merging)
